+++ bin/update_dbs.dart
@@ -7,6 +7,7 @@ import 'package:elec_server/src/db/isoexpress/ncpc_economic_report.dart';
 import 'package:elec_server/src/db/isoexpress/ncpc_generator_performance_audit_report.dart';
 import 'package:elec_server/src/db/isoexpress/ncpc_lscpr_report.dart';
 import 'package:elec_server/src/db/lib_iso_express.dart';
+import 'package:elec_server/src/db/lib_update_dbs.dart';
 import 'package:timezone/data/latest.dart';
 import 'package:timezone/standalone.dart';
 import 'package:dotenv/dotenv.dart' as dotenv;
@@ -19,17 +20,6 @@ import 'package:elec_server/src/db/isoexpress/rt_system_demand_hourly.dart';
 import 'package:elec_server/src/db/isoexpress/da_energy_offer.dart';
 import 'package:elec_server/src/db/isoexpress/da_demand_bid.dart';
 
-Future<void> updateDailyArchive(
-    DailyIsoExpressReport archive, List<Date> days) async {
-  print('Updating archive ${archive.reportName}');
-  // await archive.setupDb();
-  await archive.dbConfig.db.open();
-  for (var day in days) {
-    await archive.downloadDay(day);
-    await archive.insertDay(day);
-  }
-  await archive.dbConfig.db.close();
-}
 
 void main(List<String> args) async {
   var parser = ArgParser()